Changes for go 1.6
