+++ deferclient/client.go
@@ -8,7 +8,6 @@ import (
 	"io/ioutil"
 	"log"
 	"net/http"
-	"runtime"
 	"runtime/debug"
 	"strings"
 	"sync"
@@ -157,24 +156,6 @@ func (c *DeferPanicClient) prep(err interface{}, spanId int64, syncShipTrace boo
 	} else {
 		go c.ShipTrace(body, errorMsg, spanId)
 	}
-}
-
-// backtrace grabs the backtrace
-func backTrace() (body string) {
-
-	for skip := 1; ; skip++ {
-		pc, file, line, ok := runtime.Caller(skip)
-		if !ok {
-			break
-		}
-		if file[len(file)-1] == 'c' {
-			continue
-		}
-		f := runtime.FuncForPC(pc)
-		body += fmt.Sprintf("%s:%d %s()\n", file, line, f.Name())
-	}
-
-	return body
 }
 
 // cleanTrace should be fixed
